:recycle: Make use of LibEmitter

+++ packages/contracts/src/solecs/Emitter.sol
@@ -2,11 +2,11 @@
 pragma solidity >=0.8.0;
 
 contract Emitter {
-  event SystemCalled(uint256 indexed systemId, uint256[] schema, bytes value);
+  event SystemCalled(uint256 indexed systemId, uint8[] schema, bytes value);
 
   function emitSystemCalled(
     uint256 systemId,
-    uint256[] calldata schema,
+    uint8[] calldata schema,
     bytes calldata value
   ) external {
     emit SystemCalled(systemId, schema, value);
