+++ packages/contracts/src/solecs/interfaces/IEmitter.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0;
 interface IEmitter {
   function emitSystemCalled(
     uint256 systemId,
-    uint256[] calldata schema,
+    uint8[] calldata schema,
     bytes calldata value
   ) external;
 }
